% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as_dtype}
\alias{as_dtype}
\title{Convert to TensorDataType}
\usage{
as_dtype(x)
}
\arguments{
\item{x}{(any)\cr
Object to convert.
Can currently be a string (one of pred, i8, i16, i32, i64, ui8, ui16, ui32, ui64, f32, f64) or a \code{\link{TensorDataType}} object.}
}
\value{
\code{TensorDataType}
}
\description{
Convert to TensorDataType.
}
