% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{BoolAttr}
\alias{BoolAttr}
\title{BoolAttr}
\usage{
BoolAttr(name, value)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(\code{logical(1)})\cr
The boolean value.}
}
\value{
\code{BoolAttr}
}
\description{
An attribute holding a boolean value.
}
