% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{Constant}
\alias{Constant}
\title{Constant}
\usage{
Constant(data, type)
}
\arguments{
\item{data}{(any)\cr
The value of the constant.}

\item{type}{(\code{\link{TensorType}})\cr
The type of the constant.}
}
\value{
\code{Constant}
}
\description{
This represents a constant value.
}
