% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{constant_attr}
\alias{constant_attr}
\title{Create a ConstantAttr from R values}
\usage{
constant_attr(name, value, dtype = NULL, shape = NULL)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(any)\cr
The R value to convert to a constant.}

\item{dtype}{(\code{character(1)} | \code{NULL})\cr
The dtype of the constant. If NULL, inferred from value.}

\item{shape}{(\code{integer()} | \code{NULL})\cr
The shape of the constant. If NULL, inferred from value.}
}
\value{
(\code{ConstantAttr})
}
\description{
Helper function to create a ConstantAttr from R values.
}
