% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{ConstantAttr}
\alias{ConstantAttr}
\title{ConstantAttr}
\usage{
ConstantAttr(name = character(0), value = Constant())
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(\code{Constant})\cr
The value of the attribute.}
}
\value{
(\code{ConstantAttr})
}
\description{
An attribute holding a constant value.
}
