% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-convolution.R
\name{ConvDimensionNumbers}
\alias{ConvDimensionNumbers}
\title{ConvDimensionNumbers}
\usage{
ConvDimensionNumbers(
  input_batch_dimension,
  input_feature_dimension,
  input_spatial_dimensions,
  kernel_input_feature_dimension,
  kernel_output_feature_dimension,
  kernel_spatial_dimensions,
  output_batch_dimension,
  output_feature_dimension,
  output_spatial_dimensions
)
}
\arguments{
\item{input_batch_dimension}{(\code{integer(1)})\cr
Batch dimension of the input.}

\item{input_feature_dimension}{(\code{integer(1)})\cr
Feature dimension of the input.}

\item{input_spatial_dimensions}{(\code{integer()})\cr
Spatial dimensions of the input.}

\item{kernel_input_feature_dimension}{(\code{integer(1)})\cr
Input-feature dimension of the kernel.}

\item{kernel_output_feature_dimension}{(\code{integer(1)})\cr
Output-feature dimension of the kernel.}

\item{kernel_spatial_dimensions}{(\code{integer()})\cr
Spatial dimensions of the kernel.}

\item{output_batch_dimension}{(\code{integer(1)})\cr
Batch dimension of the output.}

\item{output_feature_dimension}{(\code{integer(1)})\cr
Feature dimension of the output.}

\item{output_spatial_dimensions}{(\code{integer()})\cr
Spatial dimensions of the output.}
}
\value{
(\code{ConvDimensionNumbers})
}
\description{
Represents the dimension numbers used by \code{\link{hlo_convolution}}.
All indices are 0-based.

For an input/output tensor of rank \code{N}, the batch dimension, the feature
dimension and the \code{N - 2} spatial dimensions partition \verb{[0, N)}. Likewise
for the kernel, the input feature, output feature and spatial dimensions
partition \verb{[0, N)}.
}
