% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-custom_call.R
\name{CustomCallConfig}
\alias{CustomCallConfig}
\title{CustomCallConfig}
\usage{
CustomCallConfig(
  call_target_name,
  api_version = 4L,
  has_side_effect = TRUE,
  backend_config = NULL
)
}
\arguments{
\item{call_target_name}{(\code{character(1)})\cr
The name of the custom function to call.}

\item{api_version}{(\code{integer(1)})\cr
The API version. Default is 4 (FFI API).}

\item{has_side_effect}{(\code{logical(1)})\cr
Whether the custom call has side effects. Default is TRUE.}

\item{backend_config}{(\code{list} | \code{NULL})\cr
Optional backend configuration as a named list.}
}
\value{
\code{CustomCallConfig}
}
\description{
Configuration for a custom call operation.
}
