% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{dimension_out_of_range_error}
\alias{dimension_out_of_range_error}
\title{DimensionOutOfRangeError}
\usage{
dimension_out_of_range_error(
  arg,
  dimension,
  ndims,
  call = sys.call(-1),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{dimension}{(\code{integer()})\cr The dimension index(es) that are out of range (0-based)}

\item{ndims}{(\code{integer(1)})\cr The number of dimensions of the tensor}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when a dimension index is outside the valid range [0, ndims)
}
