% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{DimensionUniquenessError}
\alias{DimensionUniquenessError}
\title{DimensionUniquenessError}
\usage{
DimensionUniquenessError(
  message = character(),
  call = NULL,
  .data = list(),
  arg = character(0),
  dimensions = integer(0)
)
}
\arguments{
\item{message}{(\code{character(1)})\cr Error message}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{.data}{(\code{list()})\cr Additional data to store in the condition}

\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{dimensions}{(\code{integer()})\cr The dimension indices that are not unique (0-based)}
}
\description{
Error when dimension indices are not unique
}
