% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{.current_func}
\alias{.current_func}
\title{Get the last function created}
\usage{
.current_func()
}
\value{
A \code{\link{Func}} object.
}
\description{
Get the last function created (either via \code{\link{hlo_func}} or \code{\link{local_func}}),
which is not returned yet.
}
