% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{.current_module}
\alias{.current_module}
\title{Get the current module}
\usage{
.current_module()
}
\value{
A \code{\link{Module}} object.
}
\description{
Get the current module created via \code{\link{hlo_module}} or \code{\link{local_module}}.
}
