% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dot_general.R
\name{DotDimensionNumbers}
\alias{DotDimensionNumbers}
\title{DotDimensionNumbers}
\usage{
DotDimensionNumbers(contracting_dims, batching_dims = NULL)
}
\arguments{
\item{contracting_dims}{(\code{integer()})\cr
The contracting dimensions.}

\item{batching_dims}{(\code{integer()} | \code{NULL})\cr
The batching dimensions.}
}
\description{
Represents the dot dimension numbers.
}
