% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-concatenate.R
\name{error_concatenate_shapes}
\alias{error_concatenate_shapes}
\title{ErrorConcatenateShapes}
\usage{
error_concatenate_shapes(
  dimensions,
  shapes,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{dimensions}{(\code{integer()})\cr The dimensions where shapes may differ (0-based)}

\item{shapes}{(\code{list()} of \code{Shape})\cr The input shapes}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when input shapes don't match (except at given dimensions)
}
