% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_dim_size_mismatch}
\alias{error_dim_size_mismatch}
\title{ErrorDimSizeMismatch}
\usage{
error_dim_size_mismatch(
  arg1,
  arg2,
  dim1,
  dim2,
  shape1,
  shape2,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg1}{(\code{character(1)})\cr Name of the first argument (e.g. "operand")}

\item{arg2}{(\code{character(1)})\cr Name of the second argument (e.g. "result")}

\item{dim1}{(\code{integer(1)})\cr Dimension index in arg1 (0-based)}

\item{dim2}{(\code{integer(1)})\cr Dimension index in arg2 (0-based)}

\item{shape1}{(\code{integer()})\cr Complete shape of arg1}

\item{shape2}{(\code{integer()})\cr Complete shape of arg2}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when a dimension size doesn't match the expected size at a given index
}
