% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_dimension_out_of_range}
\alias{error_dimension_out_of_range}
\title{ErrorDimensionOutOfRange}
\usage{
error_dimension_out_of_range(
  arg,
  dimension,
  dim_range,
  inclusive = c(TRUE, FALSE),
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{dimension}{(\code{integer()})\cr All dimension index(es). The error message will
identify which ones are out of range.}

\item{dim_range}{(\code{integer(2)})\cr The valid dimension range as c(min, max).}

\item{inclusive}{(\code{logical(2)})\cr Whether lower and upper bounds are inclusive, as c(lower_inclusive, upper_inclusive).
Default is c(TRUE, FALSE) meaning [lower, upper)}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when a dimension index is outside the valid range
}
