% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_dimension_uniqueness}
\alias{error_dimension_uniqueness}
\title{ErrorDimensionUniqueness}
\usage{
error_dimension_uniqueness(
  arg,
  dimensions,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{dimensions}{(\code{integer()})\cr The dimension indices that are not unique (0-based)}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when dimension indices are not unique
}
