% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_index_in_set}
\alias{error_index_in_set}
\title{ErrorIndexInSet}
\usage{
error_index_in_set(
  arg1,
  arg2,
  index,
  set,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg1}{(\code{character(1)})\cr Name of the argument containing the index}

\item{arg2}{(\code{character(1)})\cr Name of the argument containing the set}

\item{index}{(\code{integer(1)})\cr The index that was found in the set}

\item{set}{(\code{integer()})\cr The set that should not contain the index}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when an index is found in a forbidden set
}
