% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_indices_not_sorted}
\alias{error_indices_not_sorted}
\title{ErrorIndicesNotSorted}
\usage{
error_indices_not_sorted(
  arg,
  indices,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{indices}{(\code{integer()})\cr The indices that are not sorted.}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when indices are not sorted in ascending order
}
