% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_permutation}
\alias{error_permutation}
\title{Permutation Error}
\usage{
error_permutation(arg, permutation, ndims, call = NULL, signal = TRUE)
}
\arguments{
\item{arg}{Name of the argument being checked.}

\item{permutation}{The permutation values that are invalid (0-based).}

\item{ndims}{The number of dimensions of the tensor.}

\item{call}{The calling context for the error.}

\item{signal}{If TRUE (default), signals the error. If FALSE, returns the condition object.}
}
\description{
Creates and optionally signals an error when permutation values are invalid.
}
