% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_permute_index}
\alias{error_permute_index}
\title{ErrorPermuteIndex}
\usage{
error_permute_index(
  arg,
  permutation,
  expected,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{permutation}{(\code{integer()})\cr The permutation values that are invalid (0-based)}

\item{expected}{(\code{integer()})\cr The expected indices to be permuted (0-based)}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when permutation values are invalid (not a valid permutation of indices)
}
