% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_shape_mismatch}
\alias{error_shape_mismatch}
\title{Shape Mismatch Error}
\usage{
error_shape_mismatch(
  arg_lhs,
  arg_rhs,
  dim_lhs,
  dim_rhs,
  size_lhs,
  size_rhs,
  call = NULL,
  signal = TRUE
)
}
\arguments{
\item{arg_lhs}{Name of the left-hand side argument.}

\item{arg_rhs}{Name of the right-hand side argument.}

\item{dim_lhs}{Dimension index in the left-hand side tensor.}

\item{dim_rhs}{Dimension index in the right-hand side tensor.}

\item{size_lhs}{Size of the dimension in the left-hand side tensor.}

\item{size_rhs}{Size of the dimension in the right-hand side tensor.}

\item{call}{The calling context for the error.}

\item{signal}{If TRUE (default), signals the error. If FALSE, returns the condition object.}
}
\description{
Creates and optionally signals an error when dimension sizes don't match between two tensors.
}
