% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_unequal_types}
\alias{error_unequal_types}
\title{ErrorUnequalTypes}
\usage{
error_unequal_types(
  arg1,
  arg2,
  index,
  expected,
  actual1,
  actual2,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg1}{(\code{character(1)})\cr Name of the first argument}

\item{arg2}{(\code{character(1)})\cr Name of the second argument}

\item{index}{(\code{integer(1)})\cr The index where types don't match (0-based)}

\item{expected}{(\code{character(1)})\cr Description of what was expected}

\item{actual1}{(\code{character(1)})\cr Type from the first argument}

\item{actual2}{(\code{character(1)})\cr Type from the second argument}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when types at the same index in two lists don't match
}
