% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_unexpected_list_type}
\alias{error_unexpected_list_type}
\title{ErrorUnexpectedListType}
\usage{
error_unexpected_list_type(
  arg,
  index,
  expected,
  actual,
  call = sys.call(-1)[1L],
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument}

\item{index}{(\code{integer(1)})\cr The index where the type is unexpected (0-based)}

\item{expected}{(\code{character(1)})\cr Description of what was expected}

\item{actual}{(\code{character(1)})\cr What was observed}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when an element in a list has an unexpected type
}
