% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_double.R
\name{format_double}
\alias{format_double}
\title{Format Double Array with Scientific Notation}
\usage{
format_double(x, precision = 64)
}
\arguments{
\item{x}{(\code{double()})\cr
Vector to format.}

\item{precision}{(\code{integer(1)})\cr
Currently supports 32 and 64 bit precisions.}
}
\value{
\code{character()}
}
\description{
Formats a double array using scientific notation with 16 digits precision,
similar to \code{formatC(x, digits = 16, format = "e")}.

This is used to embed floating point constants into stableHLO programs.
}
\examples{
format_double(1.23, 32)
format_double(1.23, 64)
}
