% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{Func}
\alias{Func}
\title{Func}
\usage{
Func(id = FuncId(), inputs = FuncInputs(), outputs = FuncOutputs())
}
\arguments{
\item{id}{(\code{FuncId}\cr
The id of the function.}

\item{inputs}{(\code{FuncInputs}\cr
The inputs of the function.}

\item{outputs}{(\code{FuncOutputs}\cr
The outputs of the function.}
}
\value{
A \code{\link{Func}} object.
}
\description{
This represents a function.
Note: Func uses reference semantics - modifications to a Func object modify the original.
}
