% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{FuncBody}
\alias{FuncBody}
\title{FuncBody}
\usage{
FuncBody(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{Op}})\cr
The operations in the function body.}
}
\value{
(\code{FuncBody})
}
\description{
The body of a \code{\link{Func}}, containing a list of operations.
}
