% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{FuncInput}
\alias{FuncInput}
\title{FuncInput}
\usage{
FuncInput(id = ValueId(), type = ValueType(), alias = NULL)
}
\arguments{
\item{id}{(\code{\link{ValueId}})\cr
The id of the input.}

\item{type}{(\code{\link{ValueType}})\cr
The type of the input.}

\item{alias}{(\code{integer(1)} | \code{NULL})\cr
With which output buffer to alias this input.}
}
\value{
(\code{FuncInput})
}
\description{
This represents an input of a \code{\link{Func}}.
}
