% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{FuncInputs}
\alias{FuncInputs}
\title{FuncInputs}
\usage{
FuncInputs(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{FuncInput}})\cr
The inputs of the function.}
}
\value{
(\code{FuncInputs})
}
\description{
List of \code{\link{FuncInput}}s.
}
