% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{FuncOutput}
\alias{FuncOutput}
\title{FuncOutput}
\usage{
FuncOutput(type = ValueType())
}
\arguments{
\item{type}{(\code{ValueType})\cr
The type of the output.}
}
\value{
(\code{FuncOutput})
}
\description{
This represents an output of a \code{\link{Func}}.
}
