% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{FuncOutputs}
\alias{FuncOutputs}
\title{FuncOutputs}
\usage{
FuncOutputs(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{FuncOutput}})\cr
The outputs of the function.}
}
\value{
(\code{FuncOutputs})
}
\description{
List of \code{\link{FuncOutput}}s.
}
