% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-gather.R
\name{GatherDimensionNumbers}
\alias{GatherDimensionNumbers}
\title{GatherDimensionNumbers}
\usage{
GatherDimensionNumbers(
  offset_dims,
  collapsed_slice_dims,
  operand_batching_dims = integer(),
  start_indices_batching_dims = integer(),
  start_index_map,
  index_vector_dim
)
}
\arguments{
\item{offset_dims}{(\code{integer()})\cr
The offset dimensions.}

\item{collapsed_slice_dims}{(\code{integer()})\cr
The collapsed slice dimensions.}

\item{operand_batching_dims}{(\code{integer()})\cr
The operand batching dimensions.}

\item{start_indices_batching_dims}{(\code{integer()})\cr
The start indices batching dimensions.}

\item{start_index_map}{(\code{integer()})\cr
Maps start indices to operand dimensions.}

\item{index_vector_dim}{(\code{integer(1)})\cr
The index vector dimension.}
}
\description{
Represents the gather dimension numbers.
}
