% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atan2.R
\name{infer_types_atan2}
\alias{infer_types_atan2}
\alias{hlo_atan2}
\title{Atan2 Operator}
\usage{
infer_types_atan2(lhs, rhs)

hlo_atan2(lhs, rhs, output_types = NULL)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#atan2} for details.
}
