% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-broadcast_in_dim.R
\name{infer_types_broadcast_in_dim}
\alias{infer_types_broadcast_in_dim}
\alias{hlo_broadcast_in_dim}
\title{BroadcastInDim Operator}
\usage{
infer_types_broadcast_in_dim(operand, broadcast_dimensions, shape)

hlo_broadcast_in_dim(operand, broadcast_dimensions, shape, output_types = NULL)
}
\arguments{
\item{operand, broadcast_dimensions, shape}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#broadcast_in_dim} for details.
}
