% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-case.R
\name{hlo_case}
\alias{hlo_case}
\title{Case Operator}
\usage{
hlo_case(index, ...)
}
\arguments{
\item{index, ...}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#case} for details.
}
