% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-ceil.R
\name{infer_types_ceil}
\alias{infer_types_ceil}
\alias{hlo_ceil}
\title{Ceil Operator}
\usage{
infer_types_ceil(operand)

hlo_ceil(operand, output_types = NULL)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#ceil} for details.
}
