% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlo.R
\name{hlo_closure}
\alias{hlo_closure}
\title{Create a Closure}
\usage{
hlo_closure(...)
}
\arguments{
\item{...}{(\code{\link{FuncValue}})\cr
The variables to capture.}
}
\value{
(\code{list()} of \code{\link{FuncValue}})
}
\description{
Creates a new function without any arguments that captures the provided variables.
}
\examples{
func <- local_func()
x <- hlo_input("x", "f32", shape = c(2, 2))
y <- hlo_input("y", "f32", shape = c(2, 2))
f <- hlo_closure(x, y)
print(f)
}
