% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-concatenate.R
\name{infer_types_concatenate}
\alias{infer_types_concatenate}
\alias{hlo_concatenate}
\title{Concatenate Operator}
\usage{
infer_types_concatenate(..., dimension)

hlo_concatenate(..., dimension, output_types = NULL)
}
\arguments{
\item{..., dimension}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#concatenate} for details.
}
