% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-constant.R
\name{hlo_constant}
\alias{hlo_constant}
\alias{hlo_scalar}
\alias{hlo_scalar.logical}
\alias{hlo_tensor}
\title{Create a Constant}
\usage{
hlo_scalar(value, elt_type = NULL, ...)

\method{hlo_scalar}{logical}(value, elt_type = NULL, ...)

hlo_tensor(value, elt_type = NULL, ...)
}
\arguments{
\item{value}{(any)\cr
Value from which to create a constant.}

\item{elt_type}{(\code{character(1)})\cr
String for element type.
Can be one of f64, f32, u8, u16, u32, u64, i8, i16, i32, i64, pred.}

\item{...}{(any)\cr
Additional arguments.}
}
\description{
Create either a scalar or tensor constant.
}
\examples{
hlo_scalar(1L)
hlo_scalar(1, "f32")
hlo_tensor(array(c(1, 2, 3, 4), dim = c(1, 4)), "f32")
}
