% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dynamic_slice.R
\name{infer_types_dynamic_slice}
\alias{infer_types_dynamic_slice}
\alias{hlo_dynamic_slice}
\title{DynamicSlice Operator}
\usage{
infer_types_dynamic_slice(operand, ..., slice_sizes)

hlo_dynamic_slice(operand, ..., slice_sizes, output_types = NULL)
}
\arguments{
\item{operand, ..., slice_sizes}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#dynamic_slice} for details.
}
