% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-exponential_minus_one.R
\name{infer_types_exponential_minus_one}
\alias{infer_types_exponential_minus_one}
\alias{hlo_exponential_minus_one}
\title{ExponentialMinusOne Operator}
\usage{
infer_types_exponential_minus_one(operand)

hlo_exponential_minus_one(operand, output_types = NULL)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#exponential_minus_one} for details.
}
