% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-fft.R
\name{infer_types_fft}
\alias{infer_types_fft}
\alias{hlo_fft}
\title{Fft Operator}
\usage{
infer_types_fft(operand, fft_type, fft_length)

hlo_fft(operand, fft_type = c("FFT", "IFFT", "RFFT", "IRFFT"), fft_length)
}
\arguments{
\item{operand, fft_type, fft_length}{(\code{\link{FuncValue}})\cr}

\item{fft_type}{(\code{character(1)})\cr
One of \code{"FFT"}, \code{"IFFT"}, \code{"RFFT"}, \code{"IRFFT"}.}

\item{fft_length}{(\code{integer()})\cr
Length 1, 2, or 3 vector of \code{i64} values giving the FFT lengths along
the trailing dimensions.}

\item{operand}{(\code{FuncValue})\cr
Tensor of floating-point or complex type.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#fft} for details.
}
