% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlo.R
\name{hlo_input}
\alias{hlo_input}
\title{Create a input to a function}
\usage{
hlo_input(argname, type, shape = integer(), func_id = FuncId())
}
\arguments{
\item{argname}{(\code{character()})\cr}

\item{type}{(\code{\link{ValueType}})\cr}

\item{shape}{(\code{integer()})\cr}

\item{func_id}{(\code{\link{FuncId}})\cr}
}
\description{
Create a input to a function
}
