% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{hlo_module}
\alias{hlo_module}
\alias{local_module}
\title{Create a module}
\usage{
hlo_module()

local_module(envir = parent.frame())
}
\arguments{
\item{envir}{(\code{environment})\cr
Environment where exit handler will be registered for cleaning up the
\code{\link{Module}} if it was not finalized yet.}
}
\value{
A \code{\link{Module}} object.
}
\description{
Both functions create a new \code{\link{Module}} which is afterwards accessible via \code{\link[=.current_module]{.current_module()}}.
Functions created with \code{\link{hlo_func}} or \code{\link{local_func}} will automatically register into the
current module.
The module is finalized when a function named \code{"main"} is returned via \code{\link{hlo_return}}.

Differences between the two functions:
\itemize{
\item \code{\link{local_module}} removes the module when exiting the current scope, whereas \code{\link{hlo_module}} does not.
\item \code{\link{hlo_module}} discards the previously built module(s), whereas \code{\link{local_module}} does not.
}
}
