% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-pad.R
\name{infer_types_pad}
\alias{infer_types_pad}
\alias{hlo_pad}
\title{Pad Operator}
\usage{
infer_types_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding
)

hlo_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding,
  output_types = NULL
)
}
\arguments{
\item{operand, padding_value, edge_padding_low, edge_padding_high, interior_padding}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#pad} for details.
}
