% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reduce.R
\name{hlo_reduce}
\alias{hlo_reduce}
\title{Reduce Operator}
\usage{
hlo_reduce(inputs, init_values, dimensions, body)
}
\arguments{
\item{inputs, init_values, dimensions, body}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#reduce} for details.
}
