% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-rsqrt.R
\name{infer_types_rsqrt}
\alias{infer_types_rsqrt}
\alias{hlo_rsqrt}
\title{Rsqrt Operator}
\usage{
infer_types_rsqrt(operand)

hlo_rsqrt(operand, output_types = NULL)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#rsqrt} for details.
}
