% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-scatter.R
\name{infer_types_scatter}
\alias{infer_types_scatter}
\alias{hlo_scatter}
\title{Scatter Operator}
\usage{
infer_types_scatter(
  inputs,
  scatter_indices,
  updates,
  scatter_dimension_numbers,
  indices_are_sorted,
  unique_indices,
  update_computation
)

hlo_scatter(
  inputs,
  scatter_indices,
  updates,
  scatter_dimension_numbers,
  indices_are_sorted = FALSE,
  unique_indices = FALSE,
  update_computation
)
}
\arguments{
\item{inputs, scatter_indices, updates}{(\code{\link{FuncValue}})\cr}

\item{scatter_dimension_numbers}{(\code{ScatterDimensionNumbers})\cr
The scatter dimension numbers.}

\item{indices_are_sorted}{(\code{logical(1)})\cr
Whether indices are sorted.}

\item{unique_indices}{(\code{logical(1)})\cr
Whether indices are unique.}

\item{update_computation}{(\code{\link{Func}})\cr
The update computation function.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#scatter} for details.
}
