% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-select.R
\name{infer_types_select}
\alias{infer_types_select}
\alias{hlo_select}
\title{Select Operator}
\usage{
infer_types_select(pred, on_true, on_false)

hlo_select(pred, on_true, on_false, output_types = NULL)
}
\arguments{
\item{pred, on_true, on_false}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#select} for details.
}
