% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-shift_right_arithmetic.R
\name{infer_types_shift_right_arithmetic}
\alias{infer_types_shift_right_arithmetic}
\alias{hlo_shift_right_arithmetic}
\title{ShiftRightArithmetic Operator}
\usage{
infer_types_shift_right_arithmetic(lhs, rhs)

hlo_shift_right_arithmetic(lhs, rhs, output_types = NULL)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#shift_right_arithmetic} for details.
}
