% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sort.R
\name{hlo_sort}
\alias{hlo_sort}
\title{mnemonic sort}
\usage{
hlo_sort(..., dimension, is_stable, comparator)
}
\arguments{
\item{...}{(\code{\link{FuncVariable}})\cr}

\item{dimension}{dimension to sort on}

\item{is_stable}{boolean constant}

\item{comparator}{function}
}
\description{
mnemonic sort
}
