% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-tan.R
\name{infer_types_tan}
\alias{infer_types_tan}
\alias{hlo_tan}
\title{Tan Operator}
\usage{
infer_types_tan(operand)

hlo_tan(operand, output_types = NULL)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#tan} for details.
}
