% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-transpose.R
\name{infer_types_transpose}
\alias{infer_types_transpose}
\alias{hlo_transpose}
\title{Transpose Operator}
\usage{
infer_types_transpose(operand, permutation)

hlo_transpose(operand, permutation, output_types = NULL)
}
\arguments{
\item{operand, permutation}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#transpose} for details.
}
