% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{index_vec}
\alias{index_vec}
\title{index_vec}
\usage{
index_vec(x)
}
\arguments{
\item{x}{(\code{integer()})\cr Integer vector of 0-based indices.}
}
\value{
An integer vector with additional class \code{"index_vec"}.
}
\description{
Wraps an integer vector marking it as containing 0-based index values.
}
