% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_inference.R
\name{infer_types_generic_biv}
\alias{infer_types_generic_biv}
\title{Infer types for binary operations}
\usage{
infer_types_generic_biv(lhs, rhs)
}
\arguments{
\item{lhs}{(\code{ValueType})\cr
The left-hand side operand.}

\item{rhs}{(\code{ValueType})\cr
The right-hand side operand.}
}
\value{
(\code{ValueType})\cr
The inferred type. Where one operand has a dynamic axis and the other a
known size, the known size wins.
}
\description{
Infer the types for binary operations.
}
