% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_inference.R
\name{infer_types_integerish_biv}
\alias{infer_types_integerish_biv}
\title{Infer types for boolean integerish operations}
\usage{
infer_types_integerish_biv(lhs, rhs)
}
\arguments{
\item{lhs}{(\code{ValueType})\cr
The left-hand side operand.}

\item{rhs}{(\code{ValueType})\cr
The right-hand side operand.}
}
\value{
(\code{ValueType})\cr
The inferred type.
}
\description{
Infer the types for integerish (bool or int) binary operations.
}
