% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_inference.R
\name{infer_types_integerish_uni}
\alias{infer_types_integerish_uni}
\title{Infer types for integerish unary operations}
\usage{
infer_types_integerish_uni(operand)
}
\arguments{
\item{operand}{(\code{ValueType})\cr
The operand.}
}
\value{
(\code{ValueType})\cr
The inferred type.
}
\description{
Infer the types for integerish (bool or int) unary operations.
}
