% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{InferenceError}
\alias{InferenceError}
\title{InferenceError}
\usage{
InferenceError(message = character(), call = NULL, ...)
}
\arguments{
\item{message}{(\code{character(1)})\cr Error message}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{...}{Additional fields to store in the condition}
}
\description{
Base class for type inference errors
}
