% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{local_func}
\alias{local_func}
\title{Create a function}
\usage{
local_func()
}
\arguments{
\item{id}{(\code{character(1)}\cr
The id of the function.}
}
\value{
A \code{\link{Func}} object.
}
\description{
Create a local function.
}
