% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{make_error_function}
\alias{make_error_function}
\title{Factory function to create error helper functions}
\usage{
make_error_function(error_class)
}
\arguments{
\item{error_class}{The S7 error class}
}
\value{
A function that creates and throws the error
}
\description{
Factory function to create error helper functions
}
