% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{Module}
\alias{Module}
\title{Module}
\usage{
Module(funcs = list())
}
\arguments{
\item{funcs}{(\code{list()} of \code{\link{Func}})\cr
The functions in the module.}
}
\value{
A \code{Module} object.
}
\description{
A StableHLO module containing one or more \code{\link{Func}} objects.
Modules allow defining multiple named functions, where functions can
call each other using \code{\link{hlo_call}}.

Note: Module uses reference semantics like \code{\link{Func}}.
}
