% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{new_enum}
\alias{new_enum}
\title{Create a new enum class}
\usage{
new_enum(class_name, variants)
}
\arguments{
\item{class_name}{Name of the enum class}

\item{variants}{Character vector of valid variants}
}
\value{
Constructor function for the enum
}
\description{
Create a new enum class
}
\keyword{internal}
