% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of.R
\name{new_list_of}
\alias{new_list_of}
\title{Create a new list_of class}
\usage{
new_list_of(class_name, item_class, validator = NULL)
}
\arguments{
\item{class_name}{Name of the new class}

\item{item_class}{Name of the class items should be (for documentation)}

\item{validator}{Optional function to validate items}
}
\value{
Constructor function for the list_of subclass
}
\description{
Create a new list_of class
}
\keyword{internal}
