% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{Op}
\alias{Op}
\title{Op}
\usage{
Op(
  name = OpName(),
  inputs = OpInputs(),
  outputs = OpOutputs(),
  signature = OpSignature()
)
}
\arguments{
\item{name}{(\code{OpName})\cr
The name of the operation.}

\item{inputs}{(\code{OpInputs})\cr
The inputs to the operation.}

\item{outputs}{(\code{OpOutputs})\cr
The outputs of the operation.}

\item{signature}{(\code{OpSignature})\cr
The signature of the operation.}
}
\value{
(\code{Op})
}
\description{
This represents a StableHLO operation.
}
