% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpInputAttrs}
\alias{OpInputAttrs}
\title{OpInputAttrs}
\usage{
OpInputAttrs(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{OpInputAttr}})\cr
The attributes that can be used as inputs to operations.}
}
\value{
(\code{OpInputAttrs})
}
\description{
List of \code{\link{OpInputAttr}}s.
}
