% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{OpInputFuncs}
\alias{OpInputFuncs}
\title{OpInputFuncs}
\usage{
OpInputFuncs(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{OpInputFunc}})\cr
The functions that can be used as inputs to operations.}
}
\value{
(\code{OpInputFuncs})
}
\description{
List of \code{\link{OpInputFunc}}s.
}
