% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpInputValues}
\alias{OpInputValues}
\title{OpInputValues}
\usage{
OpInputValues(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{OpInputValue}})\cr
The values that can be used as inputs to operations.}
}
\value{
(\code{OpInputValues})
}
\description{
List of \code{\link{OpInputValue}}s.
}
