% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpOutput}
\alias{OpOutput}
\title{OpOutput}
\usage{
OpOutput(value_id = ValueId())
}
\arguments{
\item{id}{(\code{ValueId})\cr
The id of the output.}
}
\value{
(\code{OpOutput})
}
\description{
This represents an output of an operation.
}
