% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpOutputs}
\alias{OpOutputs}
\title{OpOutputs}
\usage{
OpOutputs(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{OpOutput}})\cr
The outputs of an operation.}
}
\value{
(\code{OpOutputs})
}
\description{
List of \code{\link{OpOutput}}s.
}
