% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpSignature}
\alias{OpSignature}
\title{OpSignature}
\usage{
OpSignature(input_types = ValueTypes(), output_types = ValueTypes())
}
\arguments{
\item{input_types}{(\code{\link{ValueTypes}})\cr
The types of the inputs.}

\item{output_types}{(\code{\link{ValueTypes}})\cr
The types of the outputs.}
}
\value{
(\code{OpSignature})
}
\description{
This represents the signature of an operation, defining its input and output types.
}
