% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{pascal_to_snake}
\alias{pascal_to_snake}
\title{Convert PascalCase to snake_case}
\usage{
pascal_to_snake(str)
}
\arguments{
\item{str}{Character string in PascalCase}
}
\value{
Character string in snake_case
}
\description{
Convert PascalCase to snake_case
}
