% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{r_to_constant}
\alias{r_to_constant}
\title{Convert R value to Constant}
\usage{
r_to_constant(value, dtype = NULL, shape, ...)
}
\arguments{
\item{value}{(any)\cr
The value to convert.}

\item{dtype}{(\code{character(1)})\cr
The dtype of the constant.}

\item{shape}{(\code{integer()})\cr
The shape of the constant.}

\item{...}{(any)
Additional arguments.}
}
\description{
Convert R value to Constant.
}
