% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_to_stablehlo_string}
\alias{r_to_stablehlo_string}
\title{Convert R value to StableHLO string representation}
\usage{
r_to_stablehlo_string(value, dtype)
}
\arguments{
\item{value}{The R value to convert}

\item{dtype}{(\code{character(1)})\cr
The dtype of the value}
}
\value{
A string representation suitable for StableHLO dense format
}
\description{
Convert R value to StableHLO string representation
}
