% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{is_dtype}
\alias{as_dtype}
\alias{shape}
\alias{dtype}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{xlamisc}{\code{\link[xlamisc:as_dtype]{as_dtype()}}, \code{\link[xlamisc:dtype]{dtype()}}, \code{\link[xlamisc:is_dtype]{is_dtype()}}, \code{\link[xlamisc:shape]{shape()}}}
}}

