% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_id.R
\name{repr_env2name}
\alias{repr_env2name}
\title{Convert environment to name}
\usage{
repr_env2name(x)
}
\arguments{
\item{x}{The environment to convert to a name.}
}
\value{
\code{character(1)}
}
\description{
This function converts an environment to a name.
Can only be called within a \code{repr} call.
}
