% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{ScalarAttr}
\alias{ScalarAttr}
\title{ScalarAttr}
\usage{
ScalarAttr(name, value, dtype)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(\code{numeric(1)} or \code{logical(1)})\cr
The scalar value.}

\item{dtype}{(\code{\link[xlamisc:DataType]{DataType}})\cr
The dtype of the scalar (e.g., \code{as_dtype("i32")}, \code{as_dtype("f32")}, \code{as_dtype("bool")}).}
}
\value{
\code{ScalarAttr}
}
\description{
An attribute holding a scalar value with an associated dtype.
}
