% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-scatter.R
\name{ScatterDimensionNumbers}
\alias{ScatterDimensionNumbers}
\title{ScatterDimensionNumbers}
\usage{
ScatterDimensionNumbers(
  update_window_dims,
  inserted_window_dims,
  input_batching_dims = integer(),
  scatter_indices_batching_dims = integer(),
  scatter_dims_to_operand_dims,
  index_vector_dim
)
}
\arguments{
\item{update_window_dims}{(\code{integer()})\cr
The update window dimensions.}

\item{inserted_window_dims}{(\code{integer()})\cr
The inserted window dimensions.}

\item{input_batching_dims}{(\code{integer()})\cr
The input batching dimensions.}

\item{scatter_indices_batching_dims}{(\code{integer()})\cr
The scatter indices batching dimensions.}

\item{scatter_dims_to_operand_dims}{(\code{integer()})\cr
Maps scatter dimensions to operand dimensions.}

\item{index_vector_dim}{(\code{integer(1)})\cr
The index vector dimension.}
}
\description{
Represents the scatter dimension numbers.
}
