% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{StringAttr}
\alias{StringAttr}
\title{StringAttr}
\usage{
StringAttr(name = character(0), value = character(0))
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(\code{character(1)})\cr
The string value.}
}
\value{
\code{StringAttr}
}
\description{
An attribute holding a string value.
}
