% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{TensorDataType}
\alias{TensorDataType}
\title{TensorDataType}
\format{
An object of class \code{S7_union} of length 1.
}
\usage{
TensorDataType
}
\description{
Type union of all possible data types.
}
\keyword{datasets}
