% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TensorElementType}
\alias{TensorElementType}
\title{TensorElementType}
\usage{
TensorElementType(type = BooleanType())
}
\arguments{
\item{type}{(\code{\link{BooleanType}} | \code{\link{IntegerType}} | \code{\link{FloatType}})}
}
\value{
\code{TensorElementType}
}
\description{
Represents the element type of a tensor.
}
