% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{to_one_based}
\alias{to_one_based}
\title{Convert 0-based indices to 1-based}
\usage{
to_one_based(x, ...)
}
\arguments{
\item{x}{Condition object with index_vec fields.}

\item{...}{Additional arguments (not used).}
}
\value{
Condition object with index_vec fields incremented by 1.
}
\description{
Converts all index_vec fields in a condition object from
0-based to 1-based.
}
