% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{unequal_tensor_types_error}
\alias{unequal_tensor_types_error}
\title{Unequal Tensor Types Error}
\usage{
unequal_tensor_types_error(args, call = NULL)
}
\arguments{
\item{args}{Named list of tensor types that should be equal.}

\item{call}{The calling context for the error.}
}
\description{
Throws an error when tensors have different types but are expected to match.
}
