% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{UnsignedType}
\alias{UnsignedType}
\title{UnsignedType}
\usage{
UnsignedType(value)
}
\arguments{
\item{value}{(\code{integer(1)})}
}
\value{
\code{UnsignedType}
}
\description{
Represents an unsigned integer type with a given bit width.
}
