% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ValueType}
\alias{ValueType}
\title{ValueType}
\usage{
ValueType(type, shape = NULL)
}
\arguments{
\item{type}{The type of the value (TensorType or TokenType).}

\item{shape}{The shape of the value (only used when type is character).}
}
\description{
This represents the type of a value.
}
