% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ValueTypes}
\alias{ValueTypes}
\title{ValueTypes}
\usage{
ValueTypes(items = list())
}
\arguments{
\item{items}{(\code{list()} of \code{\link{ValueType}})\cr
The types of the values.}
}
\value{
\code{ValueTypes}
}
\description{
List of \code{\link{ValueType}}s.
}
