% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{Constant}
\alias{Constant}
\title{Constant}
\usage{
Constant(value = TensorConstant())
}
\arguments{
\item{value}{(\code{\link{TensorConstant}})\cr
The value of the constant.}
}
\value{
\code{Constant}
}
\description{
This represents a constant value.
}
