% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{ConstantAttr}
\alias{ConstantAttr}
\title{ConstantAttr}
\usage{
ConstantAttr(name, value, simplify_dense = TRUE)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(\code{Constant})\cr
The value of the attribute.}

\item{simplify_dense}{(\code{logical(1)})\cr
Whether to simplify dense representation. Set to \code{FALSE} for multi-dimensional arrays.}
}
\value{
(\code{ConstantAttr})
}
\description{
An attribute holding a constant value.
}
