% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{DimensionOutOfRangeError}
\alias{DimensionOutOfRangeError}
\title{DimensionOutOfRangeError}
\usage{
DimensionOutOfRangeError(
  arg,
  dimension,
  ndims,
  message = character(),
  call = NULL,
  ...
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{dimension}{(\code{integer()})\cr The dimension index(es) that are out of range (0-based)}

\item{ndims}{(\code{integer(1)})\cr The number of dimensions of the tensor}

\item{message}{(\code{character(1)})\cr Error message}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{...}{Additional fields}
}
\description{
Error when a dimension index is outside the valid range [0, ndims)
}
