% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_dimension_out_of_range}
\alias{error_dimension_out_of_range}
\title{Dimension Out of Range Error}
\usage{
error_dimension_out_of_range(arg, dimension, ndims, call = NULL, signal = TRUE)
}
\arguments{
\item{arg}{Name of the argument being checked.}

\item{dimension}{The dimension index(es) that are out of range (0-based).}

\item{ndims}{The number of dimensions of the tensor.}

\item{call}{The calling context for the error.}

\item{signal}{If TRUE (default), signals the error. If FALSE, returns the condition object.}
}
\description{
Creates and optionally signals an error when a dimension index is outside the valid range [0, ndims).
}
