% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{error_index_out_of_bounds}
\alias{error_index_out_of_bounds}
\title{ErrorIndexOutOfBounds}
\usage{
error_index_out_of_bounds(
  arg,
  index,
  lower,
  upper,
  call = sys.call(-1),
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{arg}{(\code{character(1)})\cr Name of the argument that caused the error}

\item{index}{(\code{integer()})\cr The observed index value(s).}

\item{lower}{(\code{integer(1)})\cr Lower bound of valid range.}

\item{upper}{(\code{integer(1)})\cr Upper bound of valid range, exclusive.}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Error when an index is outside the valid range [lower, upper)
}
