% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{FuncInput}
\alias{FuncInput}
\title{FuncInput}
\usage{
FuncInput(id = ValueId(), type = ValueType())
}
\arguments{
\item{id}{(\code{ValueId})\cr
The id of the input.}
\item{type}{(\code{ValueType})\cr
The type of the input.}
}
\value{
(\code{FuncInput})
}
\description{
This represents an input of a \code{\link{Func}}.
}
