% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_variable.R
\name{FuncValue}
\alias{FuncValue}
\title{FuncValue}
\usage{
FuncValue(value_id = ValueId(), value_type = ValueType(), func = Func())
}
\arguments{
\item{value_id}{The name of the variable.}

\item{value_type}{The type of the variable.}

\item{func}{The function the variable belongs to.}
}
\description{
This represents a variable within a function.
}
