% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-abs.R
\name{hlo_abs}
\alias{hlo_abs}
\title{Abs Operator}
\usage{
hlo_abs(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#abs} for details.
}
