% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-abs.R
\name{hlo_abs}
\alias{hlo_abs}
\title{StableHLO Abs}
\usage{
hlo_abs(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec?hl=en#abs} for the semantics.
}
