% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-acos.R
\name{infer_types_acos}
\alias{infer_types_acos}
\alias{hlo_acos}
\title{Acos Operator}
\usage{
infer_types_acos(operand)

hlo_acos(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#acos} for details.
}
