% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-acosh.R
\name{infer_types_acosh}
\alias{infer_types_acosh}
\alias{hlo_acosh}
\title{Acosh Operator (CHLO)}
\usage{
infer_types_acosh(operand)

hlo_acosh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo} for the full spec.
}
