% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-add.R
\name{hlo_add}
\alias{hlo_add}
\title{StableHLO Add}
\usage{
hlo_add(lhs, rhs)
}
\arguments{
\item{lhs, }{rhs (\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec?hl=en#add} for the semantics.
}
