% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-after_all.R
\name{infer_types_after_all}
\alias{infer_types_after_all}
\alias{hlo_after_all}
\title{AfterAll Operator}
\usage{
infer_types_after_all(...)

hlo_after_all(..., output_types = NULL)
}
\arguments{
\item{...}{(\code{\link{FuncValue}})\cr}

\item{output_types}{(\code{list()} of \code{\link{ValueType}} | \code{NULL})\cr Output types known ahead of time (e.g. from type inference at trace time). When provided, type inference and its input validation are skipped.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#after_all} for details.
}
