% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-after_all.R
\name{infer_types_after_all}
\alias{infer_types_after_all}
\alias{hlo_after_all}
\title{AfterAll Operator}
\usage{
infer_types_after_all(...)

hlo_after_all(...)
}
\arguments{
\item{...}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#after_all} for details.
}
