% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-after_all.R
\name{hlo_after_all}
\alias{hlo_after_all}
\title{AfterAll Operator}
\usage{
hlo_after_all(...)
}
\arguments{
\item{...}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#after_all} for details.
}
