% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-and.R
\name{infer_types_and}
\alias{infer_types_and}
\alias{hlo_and}
\title{And Operator}
\usage{
infer_types_and(lhs, rhs)

hlo_and(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#and} for details.
}
