% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-and.R
\name{hlo_and}
\alias{hlo_and}
\title{OpAnd Operator}
\usage{
hlo_and(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#and} for details.
}
