% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-asin.R
\name{infer_types_asin}
\alias{infer_types_asin}
\alias{hlo_asin}
\title{Asin Operator}
\usage{
infer_types_asin(operand)

hlo_asin(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#asin} for details.
}
