% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-asin.R
\name{infer_types_asin}
\alias{infer_types_asin}
\alias{hlo_asin}
\title{Asin Operator (CHLO)}
\usage{
infer_types_asin(operand)

hlo_asin(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chloasin_chloasinop}
for details.
}
