% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-asinh.R
\name{infer_types_asinh}
\alias{infer_types_asinh}
\alias{hlo_asinh}
\title{Asinh Operator}
\usage{
infer_types_asinh(operand)

hlo_asinh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#asinh} for details.
}
