% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atan.R
\name{infer_types_atan}
\alias{infer_types_atan}
\alias{hlo_atan}
\title{Atan Operator (CHLO)}
\usage{
infer_types_atan(operand)

hlo_atan(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chloatan_chloatanop}
for details.
}
