% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atan.R
\name{infer_types_atan}
\alias{infer_types_atan}
\alias{hlo_atan}
\title{Atan Operator}
\usage{
infer_types_atan(operand)

hlo_atan(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#atan} for details.
}
