% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atan2.R
\name{infer_types_atan2}
\alias{infer_types_atan2}
\alias{hlo_atan2}
\title{Atan2 Operator}
\usage{
infer_types_atan2(lhs, rhs)

hlo_atan2(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#atan2} for details.
}
