% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atan2.R
\name{hlo_atan2}
\alias{hlo_atan2}
\title{StableHLO Atan2}
\usage{
hlo_atan2(lhs, rhs)
}
\arguments{
\item{lhs, }{rhs (\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec?hl=en#atan2} for the semantics.
}
