% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atanh.R
\name{infer_types_atanh}
\alias{infer_types_atanh}
\alias{hlo_atanh}
\title{Atanh Operator}
\usage{
infer_types_atanh(operand)

hlo_atanh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#atanh} for details.
}
