% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-atanh.R
\name{infer_types_atanh}
\alias{infer_types_atanh}
\alias{hlo_atanh}
\title{Atanh Operator (CHLO)}
\usage{
infer_types_atanh(operand)

hlo_atanh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chloatanh_chloatanhop}
for details.
}
