% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-bessel_i1e.R
\name{infer_types_bessel_i1e}
\alias{infer_types_bessel_i1e}
\alias{hlo_bessel_i1e}
\title{BesselI1e Operator (CHLO)}
\usage{
infer_types_bessel_i1e(operand)

hlo_bessel_i1e(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chlobessel_i1e_chlobessel_i1eop}
for details.
}
