% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-bitcast_convert.R
\name{hlo_bitcast_convert}
\alias{hlo_bitcast_convert}
\title{BitcastConvert Operator}
\usage{
hlo_bitcast_convert(operand, cast_to_dtype)
}
\arguments{
\item{operand, cast_to_dtype}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#bitcast_convert} for details.
}
