% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-bitcast_convert.R
\name{infer_types_bitcast_convert}
\alias{infer_types_bitcast_convert}
\alias{hlo_bitcast_convert}
\title{BitcastConvert Operator}
\usage{
infer_types_bitcast_convert(operand, dtype)

hlo_bitcast_convert(operand, dtype)
}
\arguments{
\item{operand, dtype}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#bitcast_convert} for details.
}
