% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-broadcast_in_dim.R
\name{infer_types_broadcast_in_dim}
\alias{infer_types_broadcast_in_dim}
\alias{hlo_broadcast_in_dim}
\title{BroadcastInDim Operator}
\usage{
infer_types_broadcast_in_dim(operand, broadcast_dimensions, shape)

hlo_broadcast_in_dim(operand, broadcast_dimensions, shape)
}
\arguments{
\item{operand, broadcast_dimensions, shape}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#broadcast_in_dim} for details.
}
