% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-broadcast_in_dim.R
\name{hlo_broadcast_in_dim}
\alias{hlo_broadcast_in_dim}
\title{BroadcastInDim Operator}
\usage{
hlo_broadcast_in_dim(operand, broadcast_dimensions, shape_out)
}
\arguments{
\item{operand, broadcast_dimensions, shape_out}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#broadcast_in_dim} for details.
}
