% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-case.R
\name{infer_types_case}
\alias{infer_types_case}
\alias{hlo_case}
\title{Case Operator}
\usage{
infer_types_case(index, ...)

hlo_case(index, ...)
}
\arguments{
\item{index, ...}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#case} for details.
}
