% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-cbrt.R
\name{infer_types_cbrt}
\alias{infer_types_cbrt}
\alias{hlo_cbrt}
\title{Cbrt Operator}
\usage{
infer_types_cbrt(operand)

hlo_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#cbrt} for details.
}
