% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-cbrt.R
\name{hlo_cbrt}
\alias{hlo_cbrt}
\title{OpCbrt Operator}
\usage{
hlo_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#cbrt} for details.
}
