% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-ceil.R
\name{hlo_ceil}
\alias{hlo_ceil}
\title{OpCeil Operator}
\usage{
hlo_ceil(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#ceil} for details.
}
