% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-cholesky.R
\name{hlo_cholesky}
\alias{hlo_cholesky}
\title{Cholesky Operator}
\usage{
hlo_cholesky(a, lower)
}
\arguments{
\item{a, lower}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#cholesky} for details.
}
