% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-clamp.R
\name{infer_types_clamp}
\alias{infer_types_clamp}
\alias{hlo_clamp}
\title{Clamp Operator}
\usage{
infer_types_clamp(min, operand, max)

hlo_clamp(min, operand, max)
}
\arguments{
\item{min, operand, max}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#clamp} for details.
}
