% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-clamp.R
\name{infer_types_clamp}
\alias{infer_types_clamp}
\alias{hlo_clamp}
\title{Clamp Operator}
\usage{
infer_types_clamp(Min, operand, Max)

hlo_clamp(Min, operand, Max)
}
\arguments{
\item{Min, operand, Max}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#clamp} for details.
}
