% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-compare.R
\name{hlo_compare}
\alias{hlo_compare}
\title{Compare Operator}
\usage{
hlo_compare(lhs, rhs, comparison_direction, compare_type)
}
\arguments{
\item{lhs, rhs, comparison_direction, compare_type}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#compare} for details.
}
