% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-compare.R
\name{infer_types_compare}
\alias{infer_types_compare}
\alias{hlo_compare}
\title{Compare Operator}
\usage{
infer_types_compare(lhs, rhs, comparison_direction, compare_type)

hlo_compare(lhs, rhs, comparison_direction, compare_type)
}
\arguments{
\item{lhs, rhs, comparison_direction, compare_type}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#compare} for details.
}
