% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-concatenate.R
\name{hlo_concatenate}
\alias{hlo_concatenate}
\title{mnemonic concatenate}
\usage{
hlo_concatenate(..., dimension)
}
\arguments{
\item{...}{(\code{\link{FuncVariable}})\cr}
}
\description{
mnemonic concatenate
}
