% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-concatenate.R
\name{hlo_concatenate}
\alias{hlo_concatenate}
\title{Concatenate Operator}
\usage{
hlo_concatenate(..., dimension)
}
\arguments{
\item{..., dimension}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#concatenate} for details.
}
