% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-constant.R
\name{hlo_constant}
\alias{hlo_constant}
\alias{hlo_scalar}
\alias{hlo_scalar.logical}
\title{Create a Constant}
\usage{
hlo_scalar(value, elt_type = NULL, ...)

\method{hlo_scalar}{logical}(value, elt_type = NULL, ...)
}
\description{
Create either a scalar or tensor constant.
}
\examples{
hlo_scalar(1L)
hlo_tensor(1L, "i16")
hlo_tensor(array(c(1, 2, 3, 4), dim = c(1, 4)), "f64")
}
