% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-constant.R
\name{hlo_constant}
\alias{hlo_constant}
\alias{hlo_scalar}
\alias{hlo_scalar.logical}
\alias{hlo_scalar.double}
\alias{hlo_scalar.integer}
\alias{hlo_tensor}
\alias{hlo_tensor.array}
\alias{hlo_tensor.integer}
\alias{hlo_tensor.logical}
\alias{hlo_tensor.double}
\title{Create a Constant}
\usage{
hlo_scalar(value, dtype = NULL, ...)

\method{hlo_scalar}{logical}(value, dtype = NULL, ...)

\method{hlo_scalar}{double}(value, dtype = NULL, ...)

\method{hlo_scalar}{integer}(value, dtype = NULL, ...)

hlo_tensor(value, dtype = NULL, ...)

\method{hlo_tensor}{array}(value, dtype = NULL, ...)

\method{hlo_tensor}{integer}(value, dtype = NULL, shape = get_dims(value), ...)

\method{hlo_tensor}{logical}(value, dtype = NULL, shape = get_dims(value), ...)

\method{hlo_tensor}{double}(value, dtype = NULL, shape = get_dims(value), ...)
}
\arguments{
\item{value}{(any)\cr
Value from which to create a constant.}

\item{dtype}{(\code{character(1)})\cr
String for element type.
Can be one of f64, f32, u8, u16, u32, u64, i8, i16, i32, i64, pred.}

\item{...}{(any)\cr
Additional arguments.}

\item{shape}{(\code{integer()})\cr
Shape of the tensor.
If not specified, the shape is inferred from the data.}
}
\description{
Create either a scalar or tensor constant.
Note that strictly speaking, stableHLO 'scalars' are simply tensors with 0 dimensions.
}
\examples{
hlo_scalar(1L, "i32")
hlo_scalar(1, "f32")
hlo_scalar(TRUE)
hlo_tensor(array(c(1, 2, 3, 4), dim = c(1, 4)), "f32")
}
