% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-constant.R
\name{hlo_constant}
\alias{hlo_constant}
\alias{hlo_scalar}
\alias{hlo_tensor}
\alias{hlo_empty}
\title{Create a Constant}
\usage{
hlo_scalar(value, ..., func = .current_func())

hlo_tensor(value, ..., func = .current_func())

hlo_empty(dtype, shape, func = .current_func())
}
\arguments{
\item{value}{(any)\cr
Value from which to create a constant.}

\item{...}{(any)\cr
Additional arguments.}

\item{func}{(\code{\link{Func}})\cr
The function to add the constant to.
Per default, uses the last function created with \code{\link{hlo_func}} or \code{\link{local_func}}.}

\item{dtype}{(\code{character(1)})\cr
One of: pred, i8, i16, i32, i64, ui8, ui16, ui32, ui64, f32, f64.}

\item{shape}{(\code{integer()})\cr
The shape}
}
\description{
Create either a "scalar" (\code{\link{hlo_scalar}}) or tensor (\code{\link{hlo_tensor}}) constant.
Strictly speaking, stableHLO "scalars" are simply tensors with 0 dimensions.
To create an empty constant (at least one dimension is 0), use \code{\link{hlo_empty}}.
}
\examples{
hlo_scalar(1L, dtype = "i32", func = Func())
hlo_scalar(1, dtype = "f32", func = Func())
hlo_scalar(TRUE, func = Func())
hlo_tensor(array(c(1, 2, 3, 4), dim = c(1, 4)), dtype = "f32", func = Func())
hlo_empty(dtype = "f32", shape = c(0, 3), func = Func())
}
