% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-constant.R
\name{hlo_constant}
\alias{hlo_constant}
\alias{hlo_scalar}
\alias{hlo_scalar.logical}
\alias{hlo_scalar.double}
\alias{hlo_scalar.integer}
\alias{hlo_tensor}
\alias{hlo_tensor.array}
\alias{hlo_tensor.integer}
\alias{hlo_tensor.logical}
\alias{hlo_tensor.double}
\alias{hlo_tensor.PJRTBuffer}
\title{Create a Constant}
\usage{
hlo_scalar(x, ...)

\method{hlo_scalar}{logical}(x, ...)

\method{hlo_scalar}{double}(x, ...)

\method{hlo_scalar}{integer}(x, ...)

hlo_tensor(x, ...)

\method{hlo_tensor}{array}(x, ...)

\method{hlo_tensor}{integer}(x, ...)

\method{hlo_tensor}{logical}(x, ...)

\method{hlo_tensor}{double}(x, ...)

\method{hlo_tensor}{PJRTBuffer}(x, ...)
}
\arguments{
\item{x}{(any)\cr
Value from which to create a constant.}

\item{...}{(any)\cr
Additional arguments including:
\itemize{
\item \code{dtype} (\code{character(1)}): String for element type.
Can be one of f64, f32, u8, u16, u32, u64, i8, i16, i32, i64, pred.
\item \code{shape} (\code{integer()}): Shape of the tensor (for hlo_tensor only).
If not specified, the shape is inferred from the data.
}}
}
\description{
Create either a scalar or tensor constant.
Note that strictly speaking, stableHLO 'scalars' are simply tensors with 0 dimensions.
}
\examples{
hlo_scalar(1L, dtype = "i32")
hlo_scalar(1, dtype = "f32")
hlo_scalar(TRUE)
hlo_tensor(array(c(1, 2, 3, 4), dim = c(1, 4)), dtype = "f32")
}
