% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-convert.R
\name{infer_types_convert}
\alias{infer_types_convert}
\alias{hlo_convert}
\title{Convert Operator}
\usage{
infer_types_convert(operand, dtype)

hlo_convert(operand, dtype)
}
\arguments{
\item{operand, dtype}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#convert} for details.
}
