% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-cosh.R
\name{infer_types_cosh}
\alias{infer_types_cosh}
\alias{hlo_cosh}
\title{Cosh Operator}
\usage{
infer_types_cosh(operand)

hlo_cosh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#cosh} for details.
}
