% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-cosine.R
\name{hlo_cosine}
\alias{hlo_cosine}
\title{Cosine Operator}
\usage{
hlo_cosine(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#cosine} for details.
}
