% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-cosine.R
\name{hlo_cosine}
\alias{hlo_cosine}
\title{OpCosine Operator}
\usage{
hlo_cosine(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#cosine} for details.
}
