% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-count_leading_zeros.R
\name{hlo_count_leading_zeros}
\alias{hlo_count_leading_zeros}
\title{CountLeadingZeros Operator}
\usage{
hlo_count_leading_zeros(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#count_leading_zeros} for details.
}
