% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-digamma.R
\name{infer_types_digamma}
\alias{infer_types_digamma}
\alias{hlo_digamma}
\title{Digamma Operator}
\usage{
infer_types_digamma(operand)

hlo_digamma(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#digamma} for details.
}
