% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-divide.R
\name{hlo_divide}
\alias{hlo_divide}
\title{OpDivide Operator}
\usage{
hlo_divide(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#divide} for details.
}
