% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dot_general.R
\name{infer_types_dot_general}
\alias{infer_types_dot_general}
\alias{hlo_dot_general}
\title{DotGeneral Operator}
\usage{
infer_types_dot_general(lhs, rhs, dot_dimension_numbers)

hlo_dot_general(lhs, rhs, contracting_dims, batching_dims = NULL)
}
\arguments{
\item{lhs, rhs, contracting_dims, batching_dims}{(\code{\link{FuncValue}})\cr}

\item{dot_dimension_numbers}{(\code{DotDimensionNumbers})\cr
The dot dimension number.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#dot_general} for details.
}
