% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dot_general.R
\name{hlo_dot_general}
\alias{hlo_dot_general}
\title{DotGeneral Operator}
\usage{
hlo_dot_general(
  lhs,
  rhs,
  contracting_dims,
  batching_dims = list(integer(), integer())
)
}
\arguments{
\item{lhs, rhs, contracting_dims, batching_dims}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#dot_general} for details.
}
