% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dynamic_slice.R
\name{infer_types_dynamic_slice}
\alias{infer_types_dynamic_slice}
\alias{hlo_dynamic_slice}
\title{DynamicSlice Operator}
\usage{
infer_types_dynamic_slice(operand, ..., slice_sizes)

hlo_dynamic_slice(operand, ..., slice_sizes)
}
\arguments{
\item{operand, ..., slice_sizes}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#dynamic_slice} for details.
}
