% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dynamic_update_slice.R
\name{hlo_dynamic_update_slice}
\alias{hlo_dynamic_update_slice}
\title{DynamicUpdateSlice Operator}
\usage{
hlo_dynamic_update_slice(operand, update, start_indices)
}
\arguments{
\item{operand, update, start_indices}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#dynamic_update_slice} for details.
}
