% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-dynamic_update_slice.R
\name{infer_types_dynamic_update_slice}
\alias{infer_types_dynamic_update_slice}
\alias{hlo_dynamic_update_slice}
\title{DynamicUpdateSlice Operator}
\usage{
infer_types_dynamic_update_slice(operand, update, ...)

hlo_dynamic_update_slice(operand, update, ...)
}
\arguments{
\item{operand, update, ...}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#dynamic_update_slice} for details.
}
