% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-erf.R
\name{infer_types_erf}
\alias{infer_types_erf}
\alias{hlo_erf}
\title{Erf Operator (CHLO)}
\usage{
infer_types_erf(operand)

hlo_erf(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chloerf_chloerfop}
for details.
}
