% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-erf_inv.R
\name{infer_types_erf_inv}
\alias{infer_types_erf_inv}
\alias{hlo_erf_inv}
\title{ErfInv Operator (CHLO)}
\usage{
infer_types_erf_inv(operand)

hlo_erf_inv(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chloerf_inv_chloerf_invop}
for details.
}
