% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-erfc.R
\name{infer_types_erfc}
\alias{infer_types_erfc}
\alias{hlo_erfc}
\title{Erfc Operator (CHLO)}
\usage{
infer_types_erfc(operand)

hlo_erfc(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chloerfc_chloerfcop}
for details.
}
