% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-exponential.R
\name{hlo_exponential}
\alias{hlo_exponential}
\title{OpExponential Operator}
\usage{
hlo_exponential(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#exponential} for details.
}
