% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-exponential.R
\name{infer_types_exponential}
\alias{infer_types_exponential}
\alias{hlo_exponential}
\title{Exponential Operator}
\usage{
infer_types_exponential(operand)

hlo_exponential(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#exponential} for details.
}
