% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-exponential_minus_one.R
\name{hlo_exponential_minus_one}
\alias{hlo_exponential_minus_one}
\title{ExponentialMinusOne Operator}
\usage{
hlo_exponential_minus_one(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#exponential_minus_one} for details.
}
