% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-floor.R
\name{hlo_floor}
\alias{hlo_floor}
\title{Floor Operator}
\usage{
hlo_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#floor} for details.
}
