% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-floor.R
\name{infer_types_floor}
\alias{infer_types_floor}
\alias{hlo_floor}
\title{Floor Operator}
\usage{
infer_types_floor(operand)

hlo_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#floor} for details.
}
