% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{hlo_func}
\alias{hlo_func}
\alias{local_func}
\title{Create a function}
\usage{
hlo_func(id = "main")

local_func(id = "main")
}
\arguments{
\item{id}{(\code{character(1)}\cr
The id of the function.}
}
\value{
A \code{\link{Func}} object.
}
\description{
Create a function with the given id.
\code{\link{local_func}} removes the function when exiting the current scope, whereas
\code{\link{hlo_func}} does not.
After calling this function, the created function is stored in a global variable and accessible via \link{.current_func}.
Functions receiving a \code{\link{Func}} as an argument usually use \code{\link[=.current_func]{.current_func()}} by default.
}
