% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{hlo_func}
\alias{hlo_func}
\title{Create a function}
\usage{
hlo_func(id = "main")
}
\arguments{
\item{id}{(\code{character(1)}\cr
The id of the function.}
}
\value{
A \code{\link{Func}} object.
}
\description{
Create a function with the given id.
}
