% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-if.R
\name{infer_types_if}
\alias{infer_types_if}
\alias{hlo_if}
\title{If Operator}
\usage{
infer_types_if(pred, true_branch, false_branch)

hlo_if(pred, true_branch, false_branch, simplify = TRUE)
}
\arguments{
\item{pred, true_branch, false_branch}{(\code{\link{FuncVariable}})\cr}

\item{simplify}{(\code{logical(1)})\cr
Whether to simplify results by unpacking lists of length 1 into their single element.}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#if} for details.
}
