% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-if.R
\name{hlo_if}
\alias{hlo_if}
\title{If Operator}
\usage{
hlo_if(pred, true_branch, false_branch, simplify = TRUE)
}
\arguments{
\item{pred, true_branch, false_branch, simplify}{(\code{\link{FuncVariable}})\cr}

\item{simplify}{(\code{logical(1)})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#if} for details.
}
