% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-if.R
\name{hlo_if}
\alias{hlo_if}
\title{StableHLO If}
\usage{
hlo_if(pred, true_branch, false_branch)
}
\arguments{
\item{pred, true_branch, false_branch}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec?hl=en#if} for the semantics.
}
