% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlo.R
\name{hlo_input}
\alias{hlo_input}
\title{Create a input to a function}
\usage{
hlo_input(name, dtype, shape = integer(), func = .current_func(), alias = NULL)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the parameter.}

\item{dtype}{(\code{\link{ValueType}})\cr
The data type of the parameter.
Can contain digits, letters and underscores.
If it starts with a digit, it can only contain digits.
Otherwise it must start with a letter.}

\item{shape}{(\code{integer()})\cr
The shape of the parameter.
Use \code{integer()} for scalars.}

\item{func}{(\code{\link{Func}})\cr
The function id of the parameter.
Per default, uses the last function created with \code{\link{hlo_func}}.}

\item{alias}{(\code{integer(1)} or \code{NULL})\cr
If integer, marks this input as alias with the given output index (0-based).}
}
\description{
Create a input to a function
}
\examples{
func <- hlo_func()
x <- hlo_input("x", "f32", shape = c(2, 2))
print(x)

# You can combine multiple inputs as follows:
c(
  hlo_input("x", "f32", shape = c(2, 2)),
  hlo_input("y", "f32", shape = c(2, 2))
)
}
