% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlo.R
\name{hlo_input}
\alias{hlo_input}
\title{Create a input to a function}
\usage{
hlo_input(argname, elt_type, shape = integer(), func_id = FuncId())
}
\arguments{
\item{argname}{(\code{character()})\cr}

\item{elt_type}{(\code{\link{ValueType}})\cr}

\item{shape}{(\code{integer()})\cr}

\item{func_id}{(\code{\link{FuncId}})\cr}
}
\description{
Create a input to a function
}
\examples{
x <- hlo_input("x", "f32", shape = c(2, 2))
print(x)

# You can combine multiple inputs as follows:
c(
  hlo_input("x", "f32", shape = c(2, 2)),
  hlo_input("y", "f32", shape = c(2, 2))
)
}
