% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-iota.R
\name{infer_types_iota}
\alias{infer_types_iota}
\alias{hlo_iota}
\title{Iota Operator}
\usage{
infer_types_iota(iota_dimension, dtype, shape)

hlo_iota(iota_dimension, dtype, shape, func = NULL)
}
\arguments{
\item{iota_dimension}{(\code{integer(1)})\cr
The dimension along which to generate increasing values.
Must be in range \verb{[0, rank(output))}.}

\item{dtype}{(\code{character(1)})\cr
The data type of the output tensor.
One of: pred, i8, i16, i32, i64, ui8, ui16, ui32, ui64, f32, f64 (excluding boolean).}

\item{shape}{(\code{integer()})\cr
The shape of the output tensor.}

\item{func}{(\code{\link{Func}})\cr
The function to add the operation to.
Per default, uses the last function created with \code{\link{hlo_func}} or \code{\link{local_func}}.}
}
\value{
\code{\link{FuncValue}}
}
\description{
See \url{https://openxla.org/stablehlo/spec#iota} for details.
}
