% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-is_finite.R
\name{infer_types_is_finite}
\alias{infer_types_is_finite}
\alias{hlo_is_finite}
\title{IsFinite Operator}
\usage{
infer_types_is_finite(operand)

hlo_is_finite(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#is_finite} for details.
}
