% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-is_pos_inf.R
\name{infer_types_is_pos_inf}
\alias{infer_types_is_pos_inf}
\alias{hlo_is_pos_inf}
\title{IsPosInf Operator (CHLO)}
\usage{
infer_types_is_pos_inf(operand)

hlo_is_pos_inf(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chlois_pos_inf_chlois_pos_infop}
for details.
}
