% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-lgamma.R
\name{infer_types_lgamma}
\alias{infer_types_lgamma}
\alias{hlo_lgamma}
\title{Lgamma Operator (CHLO)}
\usage{
infer_types_lgamma(operand)

hlo_lgamma(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chlolgamma_chlolgammaop}
for details.
}
