% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-log.R
\name{infer_types_log}
\alias{infer_types_log}
\alias{hlo_log}
\title{Log Operator}
\usage{
infer_types_log(operand)

hlo_log(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#log} for details.
}
