% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-log.R
\name{hlo_log}
\alias{hlo_log}
\title{OpLog Operator}
\usage{
hlo_log(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#log} for details.
}
