% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-log_plus_one.R
\name{infer_types_log_plus_one}
\alias{infer_types_log_plus_one}
\alias{hlo_log_plus_one}
\title{LogPlusOne Operator}
\usage{
infer_types_log_plus_one(operand)

hlo_log_plus_one(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#log_plus_one} for details.
}
