% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-log_plus_one.R
\name{hlo_log_plus_one}
\alias{hlo_log_plus_one}
\title{LogPlusOne Operator}
\usage{
hlo_log_plus_one(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#log_plus_one} for details.
}
