% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-logistic.R
\name{infer_types_logistic}
\alias{infer_types_logistic}
\alias{hlo_logistic}
\title{Logistic Operator}
\usage{
infer_types_logistic(operand)

hlo_logistic(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#logistic} for details.
}
