% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-logistic.R
\name{hlo_logistic}
\alias{hlo_logistic}
\title{Logistic Operator}
\usage{
hlo_logistic(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#logistic} for details.
}
