% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-maximum.R
\name{infer_types_maximum}
\alias{infer_types_maximum}
\alias{hlo_maximum}
\title{Maximum Operator}
\usage{
infer_types_maximum(lhs, rhs)

hlo_maximum(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#maximum} for details.
}
