% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-minimum.R
\name{infer_types_minimum}
\alias{infer_types_minimum}
\alias{hlo_minimum}
\title{Minimum Operator}
\usage{
infer_types_minimum(lhs, rhs)

hlo_minimum(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#minimum} for details.
}
