% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-minimum.R
\name{hlo_minimum}
\alias{hlo_minimum}
\title{Minimum Operator}
\usage{
hlo_minimum(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#minimum} for details.
}
