% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-multiply.R
\name{hlo_multiply}
\alias{hlo_multiply}
\title{OpMultiply Operator}
\usage{
hlo_multiply(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#multiply} for details.
}
