% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-negate.R
\name{hlo_negate}
\alias{hlo_negate}
\title{Negate Operator}
\usage{
hlo_negate(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#negate} for details.
}
