% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-not.R
\name{hlo_not}
\alias{hlo_not}
\title{Not Operator}
\usage{
hlo_not(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#not} for details.
}
