% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-or.R
\name{infer_types_or}
\alias{infer_types_or}
\alias{hlo_or}
\title{Or Operator}
\usage{
infer_types_or(lhs, rhs)

hlo_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#or} for details.
}
