% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-or.R
\name{hlo_or}
\alias{hlo_or}
\title{OpOr Operator}
\usage{
hlo_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#or} for details.
}
