% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-pad.R
\name{infer_types_pad}
\alias{infer_types_pad}
\alias{hlo_pad}
\title{Pad Operator}
\usage{
infer_types_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding
)

hlo_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding
)
}
\arguments{
\item{operand, padding_value, edge_padding_low, edge_padding_high, interior_padding}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#pad} for details.
}
