% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-polygamma.R
\name{infer_types_polygamma}
\alias{infer_types_polygamma}
\alias{hlo_polygamma}
\title{Polygamma Operator}
\usage{
infer_types_polygamma(n, x)

hlo_polygamma(n, x)
}
\arguments{
\item{n, x}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#polygamma} for details.
}
