% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-popcnt.R
\name{infer_types_popcnt}
\alias{infer_types_popcnt}
\alias{hlo_popcnt}
\title{Popcnt Operator}
\usage{
infer_types_popcnt(operand)

hlo_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#popcnt} for details.
}
