% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-popcnt.R
\name{hlo_popcnt}
\alias{hlo_popcnt}
\title{Popcnt Operator}
\usage{
hlo_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#popcnt} for details.
}
