% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-power.R
\name{hlo_power}
\alias{hlo_power}
\title{OpPower Operator}
\usage{
hlo_power(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#power} for details.
}
