% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-power.R
\name{infer_types_power}
\alias{infer_types_power}
\alias{hlo_power}
\title{Power Operator}
\usage{
infer_types_power(lhs, rhs)

hlo_power(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#power} for details.
}
