% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reduce.R
\name{infer_types_reduce}
\alias{infer_types_reduce}
\alias{hlo_reduce}
\title{Reduce Operator}
\usage{
infer_types_reduce(..., body, dimensions)

hlo_reduce(inputs, init_values, dimensions, body)
}
\arguments{
\item{...}{(Inputs, Init values)}

\item{inputs, init_values, dimensions, body}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#reduce} for details.
}
