% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reduce_window.R
\name{infer_types_reduce_window}
\alias{infer_types_reduce_window}
\alias{hlo_reduce_window}
\title{ReduceWindow Operator}
\usage{
infer_types_reduce_window(
  ...,
  body,
  window_dimensions,
  window_strides,
  base_dilations,
  window_dilations,
  padding
)

hlo_reduce_window(
  inputs,
  init_values,
  window_dimensions,
  window_strides,
  base_dilations,
  window_dilations,
  padding,
  body
)
}
\arguments{
\item{...}{(Inputs, Init values)}

\item{body}{(\code{Func})\cr
The reduction function to apply to each window.}

\item{window_dimensions}{(\code{integer()})\cr
The size of the window in each dimension.}

\item{window_strides}{(\code{integer()})\cr
The stride of the window in each dimension.}

\item{base_dilations}{(\code{integer()})\cr
The dilation factor for the input tensor.}

\item{window_dilations}{(\code{integer()})\cr
The dilation factor for the window.}

\item{padding}{(\code{matrix})\cr
A matrix with shape \verb{[rank, 2]} specifying the padding before and after
each dimension.}

\item{inputs}{(\code{list()} of \code{\link{FuncValue}})\cr
The input tensor(s) to apply the reduction to.}

\item{init_values}{(\code{list()} of \code{\link{FuncValue}})\cr
The initial value(s) for the reduction. Must be 0-D tensors.}
}
\description{
See \url{https://openxla.org/stablehlo/spec#reduce_window} for details.
}
