% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-remainder.R
\name{hlo_remainder}
\alias{hlo_remainder}
\title{Remainder Operator}
\usage{
hlo_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#remainder} for details.
}
