% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reshape.R
\name{hlo_reshape}
\alias{hlo_reshape}
\title{Reshape Operator}
\usage{
hlo_reshape(operand, shape_out)
}
\arguments{
\item{operand, shape_out}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#reshape} for details.
}
