% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reshape.R
\name{infer_types_reshape}
\alias{infer_types_reshape}
\alias{hlo_reshape}
\title{Reshape Operator}
\usage{
infer_types_reshape(operand, shape)

hlo_reshape(operand, shape)
}
\arguments{
\item{operand, shape}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#reshape} for details.
}
