% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reverse.R
\name{hlo_reverse}
\alias{hlo_reverse}
\title{Reverse Operator}
\usage{
hlo_reverse(operand, dimensions)
}
\arguments{
\item{operand, dimensions}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#reverse} for details.
}
