% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-reverse.R
\name{infer_types_reverse}
\alias{infer_types_reverse}
\alias{hlo_reverse}
\title{Reverse Operator}
\usage{
infer_types_reverse(operand, dimensions)

hlo_reverse(operand, dimensions)
}
\arguments{
\item{operand, dimensions}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#reverse} for details.
}
