% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-rng_bit_generator.R
\name{infer_types_rng_bit_generator}
\alias{infer_types_rng_bit_generator}
\alias{hlo_rng_bit_generator}
\title{RngBitGenerator Operator}
\usage{
infer_types_rng_bit_generator(initial_state, rng_algorithm, dtype, shape)

hlo_rng_bit_generator(
  initial_state,
  rng_algorithm = c("DEFAULT", "THREE_FRY", "PHILOX"),
  dtype,
  shape
)
}
\arguments{
\item{initial_state, rng_algorithm, dtype, shape}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#rng_bit_generator} for details.
}
