% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-round_nearest_afz.R
\name{hlo_round_nearest_afz}
\alias{hlo_round_nearest_afz}
\title{RoundNearestAfz Operator}
\usage{
hlo_round_nearest_afz(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#round_nearest_afz} for details.
}
