% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-round_nearest_even.R
\name{hlo_round_nearest_even}
\alias{hlo_round_nearest_even}
\title{RoundNearestEven Operator}
\usage{
hlo_round_nearest_even(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#round_nearest_even} for details.
}
