% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-rsqrt.R
\name{infer_types_rsqrt}
\alias{infer_types_rsqrt}
\alias{hlo_rsqrt}
\title{Rsqrt Operator}
\usage{
infer_types_rsqrt(operand)

hlo_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#rsqrt} for details.
}
