% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-rsqrt.R
\name{hlo_rsqrt}
\alias{hlo_rsqrt}
\title{Rsqrt Operator}
\usage{
hlo_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#rsqrt} for details.
}
