% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-scatter.R
\name{hlo_scatter}
\alias{hlo_scatter}
\title{Scatter Operator}
\usage{
hlo_scatter(
  inputs,
  scatter_indices,
  updates,
  update_computation,
  update_window_dims,
  inserted_window_dims,
  input_batching_dims,
  scatter_indices_batching_dims,
  scatter_dims_to_operand_dims,
  index_vector_dim,
  indices_are_sorted = FALSE,
  unique_indices = FALSE
)
}
\arguments{
\item{inputs, scatter_indices, updates, update_computation, update_window_dims, inserted_window_dims, input_batching_dims, scatter_indices_batching_dims, scatter_dims_to_operand_dims, index_vector_dim, indices_are_sorted, unique_indices}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#scatter} for details.
}
