% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-select.R
\name{infer_types_select}
\alias{infer_types_select}
\alias{hlo_select}
\title{Select Operator}
\usage{
infer_types_select(pred, on_true, on_false)

hlo_select(pred, on_true, on_false)
}
\arguments{
\item{pred, on_true, on_false}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#select} for details.
}
