% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-select.R
\name{hlo_select}
\alias{hlo_select}
\title{Select Operator}
\usage{
hlo_select(pred, on_true, on_false)
}
\arguments{
\item{pred, on_true, on_false}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#select} for details.
}
