% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-shift_left.R
\name{infer_types_shift_left}
\alias{infer_types_shift_left}
\alias{hlo_shift_left}
\title{ShiftLeft Operator}
\usage{
infer_types_shift_left(lhs, rhs)

hlo_shift_left(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#shift_left} for details.
}
