% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-shift_left.R
\name{hlo_shift_left}
\alias{hlo_shift_left}
\title{ShiftLeft Operator}
\usage{
hlo_shift_left(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#shift_left} for details.
}
