% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-shift_right_arithmetic.R
\name{hlo_shift_right_arithmetic}
\alias{hlo_shift_right_arithmetic}
\title{ShiftRightArithmetic Operator}
\usage{
hlo_shift_right_arithmetic(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#shift_right_arithmetic} for details.
}
