% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-shift_right_logical.R
\name{infer_types_shift_right_logical}
\alias{infer_types_shift_right_logical}
\alias{hlo_shift_right_logical}
\title{ShiftRightLogical Operator}
\usage{
infer_types_shift_right_logical(lhs, rhs)

hlo_shift_right_logical(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#shift_right_logical} for details.
}
