% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sign.R
\name{hlo_sign}
\alias{hlo_sign}
\title{Sign Operator}
\usage{
hlo_sign(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sign} for details.
}
