% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sign.R
\name{infer_types_sign}
\alias{infer_types_sign}
\alias{hlo_sign}
\title{Sign Operator}
\usage{
infer_types_sign(operand)

hlo_sign(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sign} for details.
}
