% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sine.R
\name{hlo_sine}
\alias{hlo_sine}
\title{Sine Operator}
\usage{
hlo_sine(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sine} for details.
}
