% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sinh.R
\name{infer_types_sinh}
\alias{infer_types_sinh}
\alias{hlo_sinh}
\title{Sinh Operator}
\usage{
infer_types_sinh(operand)

hlo_sinh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sinh} for details.
}
