% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-slice.R
\name{infer_types_slice}
\alias{infer_types_slice}
\alias{hlo_slice}
\title{Slice Operator}
\usage{
infer_types_slice(operand, start_indices, limit_indices, strides)

hlo_slice(operand, start_indices, limit_indices, strides)
}
\arguments{
\item{operand, start_indices, limit_indices, strides}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#slice} for details.
}
