% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sort.R
\name{infer_types_sort}
\alias{infer_types_sort}
\alias{hlo_sort}
\title{Sort Operator}
\usage{
infer_types_sort(..., dimension, is_stable, comparator)

hlo_sort(..., dimension, is_stable, comparator)
}
\arguments{
\item{..., dimension, is_stable, comparator}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sort} for details.
}
