% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sort.R
\name{hlo_sort}
\alias{hlo_sort}
\title{Sort Operator}
\usage{
hlo_sort(..., dimension, is_stable, comparator)
}
\arguments{
\item{..., dimension, is_stable, comparator}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sort} for details.
}
