% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sqrt.R
\name{infer_types_sqrt}
\alias{infer_types_sqrt}
\alias{hlo_sqrt}
\title{Sqrt Operator}
\usage{
infer_types_sqrt(operand)

hlo_sqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sqrt} for details.
}
