% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-sqrt.R
\name{hlo_sqrt}
\alias{hlo_sqrt}
\title{Sqrt Operator}
\usage{
hlo_sqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#sqrt} for details.
}
