% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-square.R
\name{infer_types_square}
\alias{infer_types_square}
\alias{hlo_square}
\title{Square Operator (CHLO)}
\usage{
infer_types_square(operand)

hlo_square(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
This op is from the CHLO dialect, a higher-level companion to stableHLO
that is lowered to stableHLO during compilation. See
\url{https://openxla.org/stablehlo/generated/chlo#chlosquare_chlosquareop}
for details.
}
