% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-subtract.R
\name{hlo_subtract}
\alias{hlo_subtract}
\title{Subtraction}
\usage{
hlo_subtract(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})}
}
\value{
\code{\link{FuncVariable}}
}
\description{
Subtraction
}
