% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-subtract.R
\name{hlo_subtract}
\alias{hlo_subtract}
\title{Subtract Operator}
\usage{
hlo_subtract(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#subtract} for details.
}
