% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-tan.R
\name{hlo_tan}
\alias{hlo_tan}
\title{Tangent Operation}
\usage{
hlo_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})}
}
\value{
\code{\link{FuncVariable}}
}
\description{
Tangent Operation
}
