% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-tan.R
\name{hlo_tan}
\alias{hlo_tan}
\title{Tan Operator}
\usage{
hlo_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#tan} for details.
}
