% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-tanh.R
\name{hlo_tanh}
\alias{hlo_tanh}
\title{Tanh Operator}
\usage{
hlo_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#tanh} for details.
}
