% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-tanh.R
\name{hlo_tanh}
\alias{hlo_tanh}
\title{Tangens hyperbolicus}
\usage{
hlo_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})}
}
\value{
\code{\link{FuncVariable}}
}
\description{
Tangens hyperbolicus
}
