% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-transpose.R
\name{infer_types_transpose}
\alias{infer_types_transpose}
\alias{hlo_transpose}
\title{Transpose Operator}
\usage{
infer_types_transpose(operand, permutation)

hlo_transpose(operand, permutation)
}
\arguments{
\item{operand, permutation}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#transpose} for details.
}
