% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-transpose.R
\name{hlo_transpose}
\alias{hlo_transpose}
\title{Transpose}
\usage{
hlo_transpose(operand, permutation)
}
\arguments{
\item{operand}{(\code{\link{FuncVariable}})}
}
\value{
\code{\link{FuncVariable}}
}
\description{
Transpose
}
