% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-triangular_solve.R
\name{infer_types_triangular_solve}
\alias{infer_types_triangular_solve}
\alias{hlo_triangular_solve}
\title{TriangularSolve Operator}
\usage{
infer_types_triangular_solve(
  a,
  b,
  left_side,
  lower,
  unit_diagonal,
  transpose_a
)

hlo_triangular_solve(a, b, left_side, lower, unit_diagonal, transpose_a)
}
\arguments{
\item{a, b}{(\code{\link{FuncValue}})\cr}

\item{left_side}{(\code{logical(1)})\cr
If \code{TRUE}, solve \code{op(a) * x = b}. If \code{FALSE}, solve \code{x * op(a) = b}.}

\item{lower}{(\code{logical(1)})\cr
If \code{TRUE}, use lower triangle of \code{a}. If \code{FALSE}, use upper triangle.}

\item{unit_diagonal}{(\code{logical(1)})\cr
If \code{TRUE}, assume diagonal elements of \code{a} are 1.}

\item{transpose_a}{(\code{character(1)})\cr
One of \code{"NO_TRANSPOSE"}, \code{"TRANSPOSE"}, or \code{"ADJOINT"}.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#triangular_solve} for details.
}
