% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-while.R
\name{infer_types_while}
\alias{infer_types_while}
\alias{hlo_while}
\title{While Operator}
\usage{
infer_types_while(..., cond, body)

hlo_while(..., cond, body)
}
\arguments{
\item{..., cond, body}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#while} for details.
}
