% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-while.R
\name{infer_types_while}
\alias{infer_types_while}
\alias{hlo_while}
\title{While Operator}
\usage{
infer_types_while(..., cond, body)

hlo_while(..., cond, body, simplify = TRUE)
}
\arguments{
\item{..., cond, body}{(\code{\link{FuncValue}})\cr}

\item{simplify}{(\code{logical(1)})\cr
Whether to simplify results by unpacking lists of length 1 into their single element.}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#while} for details.
}
