% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-xor.R
\name{infer_types_xor}
\alias{infer_types_xor}
\alias{hlo_xor}
\title{Xor Operator}
\usage{
infer_types_xor(lhs, rhs)

hlo_xor(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncValue}})\cr}
}
\value{
\code{\link{FuncValue}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#xor} for details.
}
