% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-xor.R
\name{hlo_xor}
\alias{hlo_xor}
\title{OpXor Operator}
\usage{
hlo_xor(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{FuncVariable}})\cr}
}
\value{
\code{\link{FuncVariable}}\cr
}
\description{
See \url{https://openxla.org/stablehlo/spec#xor} for details.
}
