% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-custom_call.R
\name{infer_types_custom_call}
\alias{infer_types_custom_call}
\title{Infer types for custom call}
\usage{
infer_types_custom_call(...)
}
\arguments{
\item{...}{Input value types and custom call config.}
}
\value{
(\code{ValueTypes})\cr
The output types (empty for side-effect only calls).
}
\description{
Infer the output types for a custom call operation.
}
