% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{invalid_identifier_error}
\alias{invalid_identifier_error}
\title{Invalid Identifier Error}
\usage{
invalid_identifier_error(arg, call = NULL)
}
\arguments{
\item{arg}{The invalid identifier string.}

\item{call}{The calling context for the error.}
}
\description{
Throws an error when an identifier doesn't follow naming rules.
}
