% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpInputAttr}
\alias{OpInputAttr}
\title{OpInputAttr}
\usage{
OpInputAttr(name, value, dtype)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(any)\cr
The value of the attribute.}

\item{dtype}{(\code{\link{TensorDataType}})\cr
The dtype of the attribute.}
}
\value{
(\code{OpInputAttr})
}
\description{
Base class for operation input attributes.
}
