% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpInputAttr}
\alias{OpInputAttr}
\title{OpInputAttr}
\usage{
OpInputAttr(name = character(0), value = Constant())
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the attribute.}

\item{value}{(\code{Constant})\cr
The value of the attribute.}
}
\value{
(\code{OpInputAttr})
}
\description{
This represents an attribute that can be used as input to an operation.
}
