% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{OpInputFunc}
\alias{OpInputFunc}
\title{OpInputFunc}
\usage{
OpInputFunc(inputs = FuncInputs(), body = FuncBody())
}
\arguments{
\item{inputs}{(\code{FuncInputs})\cr
The inputs of the function.}

\item{body}{(\code{FuncBody})\cr
The body of the function.}
}
\value{
(\code{OpInputFunc})
}
\description{
This represents a function that can be used as input to an operation.
}
