% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpInputs}
\alias{OpInputs}
\title{OpInputs}
\usage{
OpInputs(
  values = OpInputValues(),
  funcs = OpInputFuncs(),
  attrs = OpInputAttrs(),
  custom_attrs = list()
)
}
\arguments{
\item{values}{(\code{\link{OpInputValues}})\cr
The values used as inputs.}

\item{funcs}{(\code{\link{OpInputFuncs}})\cr
The functions used as inputs.}

\item{attrs}{(\code{\link{OpInputAttrs}})\cr
The attributes used as inputs.}

\item{custom_attrs}{(\code{list})\cr
Custom attributes. Use this attributes that require custom formatting.}
}
\value{
(\code{OpInputs})
}
\description{
This represents all the inputs to an operation, including values, functions, and attributes.
}
