% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpInputValue}
\alias{OpInputValue}
\title{OpInputValue}
\usage{
OpInputValue(id)
}
\arguments{
\item{id}{(\code{\link{ValueId}})\cr
The id of the value.}
}
\value{
(\code{OpInputValue})
}
\description{
This represents a value that can be used as input to an operation.
}
