% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{OpName}
\alias{OpName}
\title{OpName}
\usage{
OpName(mnemonic)
}
\arguments{
\item{mnemonic}{(\code{\link{OpMnemonic}})\cr
The mnemonic of the operation.}
}
\value{
(\code{OpName})
}
\description{
This represents the name of an operation, containing a mnemonic.
}
