% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_to_stablehlo_string}
\alias{r_to_stablehlo_string}
\title{Represent an R value in stableHLO}
\usage{
r_to_stablehlo_string(value, dtype)
}
\arguments{
\item{value}{(any)\cr
The R value to convert}

\item{dtype}{(\code{character(1)})\cr
The element type to use.}
}
\value{
\code{character(1)}
}
\description{
Convert R value to StableHLO string representation
}
