% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{BooleanType}
\alias{IntegerType}
\alias{UIntegerType}
\alias{FloatType}
\alias{is_dtype}
\alias{as_dtype}
\alias{shape}
\alias{dtype}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tengen}{\code{\link[tengen]{as_dtype}}, \code{\link[tengen]{BooleanType}}, \code{\link[tengen]{dtype}}, \code{\link[tengen]{FloatType}}, \code{\link[tengen]{IntegerType}}, \code{\link[tengen]{is_dtype}}, \code{\link[tengen]{shape}}, \code{\link[tengen]{UIntegerType}}}
}}

