% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr.R
\name{repr}
\alias{repr}
\title{Generate string representation for object}
\usage{
repr(x, ...)
}
\arguments{
\item{x}{The object to generate a string representation of.}

\item{...}{Additional arguments passed to the method.}
}
\value{
\code{character(1)}
}
\description{
This function generates a string representation of an object.
In this package, this is primarily used to convert a \code{Func}
to its stableHLO string representation.
}
