% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stablehlo_error}
\alias{stablehlo_error}
\title{StablehloError}
\usage{
stablehlo_error(
  message = character(),
  call = sys.call(-1),
  ...,
  class = character(),
  signal = TRUE
)
}
\arguments{
\item{message}{(\code{character(1)})\cr Error message}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{...}{Additional fields to store in the condition}

\item{class}{(\code{character()})\cr Additional classes to prepend}

\item{signal}{(\code{logical(1)})\cr Whether to signal the error (default TRUE)}
}
\description{
Base error class for all stablehlo errors
}
