% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{StablehloError}
\alias{StablehloError}
\title{StablehloError}
\usage{
StablehloError(message = character(), call = NULL, .data = list())
}
\arguments{
\item{message}{(\code{character(1)})\cr Error message}

\item{call}{(\code{call} or \code{NULL})\cr Call that generated the error}

\item{.data}{(\code{list()})\cr Additional data to store in the condition}
}
\description{
Base error class for all stablehlo errors
}
