% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{TensorDataType}
\alias{TensorDataType}
\title{TensorDataType}
\format{
An object of class \code{character} of length 4.
}
\usage{
TensorDataType
}
\description{
Type union of all possible data types.
Not a class in S3, just documentation that dtype can be BooleanType, IntegerType, UnsignedType, or FloatType.
}
\keyword{datasets}
