% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{TensorElementType}
\alias{TensorElementType}
\title{TensorElementType}
\format{
An object of class \code{S7_union} of length 1.
}
\usage{
TensorElementType
}
\description{
Type union of all possible element types.
}
\keyword{datasets}
