% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TensorType}
\alias{TensorType}
\title{TensorType}
\usage{
TensorType(dtype = TensorElementType(), shape = Shape())
}
\arguments{
\item{dtype}{(\code{\link{TensorElementType}})}

\item{shape}{(\code{\link{Shape}})}
}
\value{
\code{TensorType}
}
\description{
Represents a tensor type with a specific data type and shape.
}
