% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{to_one_based}
\alias{to_one_based}
\title{Convert 0-based indices to 1-based}
\usage{
to_one_based(x, ...)
}
\arguments{
\item{x}{Condition object with indices}

\item{...}{Additional arguments (not used)}
}
\value{
Condition object with indices converted to 1-based
}
\description{
Generic function to convert 0-based indices in error conditions to 1-based
}
