% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_id.R
\name{ValueId}
\alias{ValueId}
\title{ValueId}
\usage{
ValueId(id = NULL)
}
\arguments{
\item{id}{(\code{character(1)} or \code{\link{environment}})\cr
Either a fixed name or an environment.
If using an environment (default), the name will be generated automatically
when calling \code{\link[=repr]{repr()}}, i.e. the first value id will be \verb{\%0}, the second \verb{\%1}, etc..}
}
\value{
(\code{ValueId})
}
\description{
This represents the name of a \code{\link{ValueType}}.
}
